Switch TIMER1 and TIMER2 functionality
Update all locations with TIMER1 and TIMER2. TIMER2 is now the millisecond timer, and TIMER1 is the microsecond timer. This is because TIMER2 is 8-bit so can't count the number of milliseconds we would want without doing extra looping, which defeats the purpose of an accurate microsecond timer

+++ Mech458Project/Tray.h
@@ -18,8 +18,8 @@
 typedef volatile struct Tray
 {
 	uint8_t currentPos;
-	uint8_t targetPos;
-	uint8_t lastDelay;
+	uint16_t targetPos;
+	uint16_t lastDelay;
 	
 	bool isReady;
 	ItemClass beltPos;
